+++ media/lua/client/CHC_config.lua
@@ -4,6 +4,7 @@ require 'CHC_config_apply_funcs'
 
 local utils = require('CHC_utils')
 
+local dir = utils.configDir
 local config_name = 'craft_helper_config.json'
 local mappings_name = 'CHC_mappings.json'
 
@@ -241,16 +242,24 @@ else
 end
 
 
-CHC_settings.Save = function()
-    utils.jsonutil.Save(config_name, CHC_settings.config)
+CHC_settings.Save = function(config)
+    config = config or CHC_settings.config
+    local status = pcall(utils.jsonutil.Save, config_name, config)
+    if not status then
+        -- config is corrupted, create new
+        CHC_settings.Save(init_cfg)
+    end
 end
 
 CHC_settings.Load = function()
-    local config = utils.jsonutil.Load(config_name)
+    local status, config = pcall(utils.jsonutil.Load, config_name)
 
-    if not config then
+    if not status or not config then
+        -- -- try to load old config first
+        -- local oldStatus, oldConfig = pcall(utils.jsonutil.Load, config_name)
+        -- config = oldStatus and oldConfig or init_cfg
         config = init_cfg
-        CHC_settings.Save()
+        CHC_settings.Save(config)
     end
     CHC_settings.checkConfig(config)
     CHC_settings.config = config
@@ -265,7 +274,7 @@ CHC_settings.checkConfig = function(config)
         end
     end
     if shouldReSave == true then
-        CHC_settings.Save()
+        CHC_settings.Save(config)
     end
 end
 
@@ -281,7 +290,10 @@ end
 
 CHC_settings.LoadPropsData = function()
     local status, config = pcall(utils.jsonutil.Load, mappings_name)
-    if not config or not status then
+    if not status or not config then
+        -- --try to load old config first
+        -- local oldStatus, oldConfig = pcall(utils.jsonutil.Load, mappings_name)
+        -- config = oldStatus and oldConfig or init_mappings
         config = init_mappings
         CHC_settings.SavePropsData(config)
     end