+++ media/lua/client/CHC_utils.lua
@@ -2,11 +2,30 @@ local CHC_utils = {}
 
 local lower = string.lower
 local tostring = tostring
+local format = string.format
 local len = string.len
 local sub = string.sub
 local insert = table.insert
 local contains = string.contains
 
+
+CHC_utils.chcprint = function(txt)
+    print('[CraftHelperContinued] ' .. tostring(txt))
+end
+
+---@param txt string error message
+---@param loc string? location of error
+---@param line number? line number of error
+CHC_utils.chcerror = function(txt, loc, line)
+    local msg = format('[CraftHelperContinued] %s', txt)
+    if loc then
+        msg = msg .. format(' at %s', loc)
+    end
+    if line then
+        msg = msg .. format(':%d', line)
+    end
+    error(msg)
+end
 
 CHC_utils.Deque = {}
 
@@ -26,7 +45,7 @@ function CHC_utils.Deque:new()
 
     function CHC_utils.Deque:_popr()
         local last = self.last
-        if self.first > last then error('deque empty') end
+        if self.first > last then CHC_utils.chcerror('Deque empty', 'CHC_utils.Deque:_popr') end
         local val = self.data[last]
         self.data[last] = nil
         self.last = last - 1
@@ -43,7 +62,7 @@ function CHC_utils.Deque:new()
 
     function CHC_utils.Deque:pop()
         local first = self.first
-        if first > self.last then error('deque empty') end
+        if first > self.last then CHC_utils.chcerror('Deque empty', 'CHC_utils.Deque:pop') end
         local val = self.data[first]
         self.data[first] = nil
         self.first = first + 1
@@ -176,27 +195,44 @@ end
 
 --- Return true if any value of 't' == 'val'
 --
----@param t table Table to check, numerical keys only
----@param val any Value to check
+---@param t table Table to check
+---@param val any Value to check, numerical keys only
 ---@param start? number Starting value (by default 1)
 ---@param stop? number Ending value (by default #t)
 ---@param step? number Step (by default 1)
-CHC_utils.any = function(t, val, start, stop, step)
+CHC_utils.any = function(t, val, start, stop, step, nonNum)
     start = start or 1
     stop = stop or #t
     step = step or 1
+    nonNum = nonNum or false
     if type(val) == 'table' then
         for j = 1, #val do
-            for i = start, stop, step do
-                if t[i] == val[j] then
-                    return true
+            if nonNum then
+                for _, value in pairs(t) do
+                    if value == val[j] then
+                        return true
+                    end
+                end
+            else
+                for i = start, stop, step do
+                    if t[i] == val[j] then
+                        return true
+                    end
                 end
             end
         end
     else
-        for i = start, stop, step do
-            if t[i] == val then
-                return true
+        if nonNum then
+            for _, value in pairs(t) do
+                if value == val then
+                    return true
+                end
+            end
+        else
+            for i = start, stop, step do
+                if t[i] == val then
+                    return true
+                end
             end
         end
     end
@@ -211,22 +247,25 @@ CHC_utils.startswith = function(txt, start)
     return sub(txt, 1, len(start)) == start
 end
 
-CHC_utils.chcprint = function(txt)
-    print('[CraftHelperContinued] ' .. tostring(txt))
-end
-
 function CHC_utils.empty(tab)
     for _, _ in pairs(tab) do return false; end
     return true
 end
 
+CHC_utils.configDir = "CraftHelperContinued" .. getFileSeparator()
+-- CHC_utils.cacheDir = CHC_utils.configDir .. "cache" .. getFileSeparator()
+
 local JsonUtil = require('CHC_json')
 
 CHC_utils.jsonutil = {}
 CHC_utils.jsonutil.Load = function(fname)
-    if not fname then error('filename not set') end
+    local func = 'CHC_utils.jsonutil.Load'
+    if not fname then CHC_utils.chcerror('Filename not set', func) end
     local res
     local fileReaderObj = getFileReader(fname, true)
+    if not fileReaderObj then
+        CHC_utils.chcerror(format('File not found and cannot be created (%s)', fname), func)
+    end
     local json = ''
     local line = fileReaderObj:readLine()
     while line ~= nil do
@@ -236,7 +275,11 @@ CHC_utils.jsonutil.Load = function(fname)
     fileReaderObj:close()
 
     if json and json ~= '' then
-        res = JsonUtil.Decode(json)
+        local status = true
+        status, res = pcall(JsonUtil.Decode, json)
+        if not status then
+            CHC_utils.chcerror(format('Cannot decode json (%s)', res), func)
+        end
     end
     return res
 end
@@ -244,7 +287,13 @@ end
 CHC_utils.jsonutil.Save = function(fname, data)
     if not data then return end
     local fileWriterObj = getFileWriter(fname, true, false)
-    local json = JsonUtil.Encode(data)
+    if not fileWriterObj then
+        CHC_utils.chcerror(format('Cannot write to %s', fname), 'CHC_utils.jsonutil.Save')
+    end
+    local status, json = pcall(JsonUtil.Encode, data)
+    if not status then
+        CHC_utils.chcerror(format('Cannot encode json (%s)', json), 'CHC_utils.jsonutil.Save')
+    end
     fileWriterObj:write(json)
     fileWriterObj:close()
 end